feat: page level flash translation layer
I made the page level flash translation layer. You can check the read
and write method on the `main.c` file.

To-do list is described below.

1. Add the R/W lock and mutex to support the SMP
2. Process the automatic garbage collection routine (currently, you must
    call the gc by using `ioctl`)
3. BlueDBM board based page ftl processing

+++ include/bits.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <limits.h>
 
-#define BITS_NOT_FOUND (UINT64_MAX)
+#define BITS_NOT_FOUND ((uint64_t)UINT64_MAX)
 
 #define BITS_PER_BYTE (8)
 #define BITS_PER_UINT64 (BITS_PER_BYTE * sizeof(uint64_t))
